+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - `PrimaryKeyFinder` now searches recursively and returns JSON Paths
 - Generate schema using 2020-12 spec
+- Tuple schemas are now produced using `prefixItems`
 
 ### Fixed
 - Arrays now correctly use `minItems` and `maxItems` instead of `minLength` and `maxLength`